app/requirements.txt
Django>=3.2, <4.0
psycopg2-binary>=2.8
djangorestframework
django-filter
djangorestframework-simplejwt
Pillow